+++ app/src/main/java/com/example/listacontato/model/Contato.kt
@@ -6,18 +6,18 @@ class Contato : Serializable{
 
     var id: Int? = null
     var nome: String
-    var numero: Int
+    var numero: Long
     var imageArrayByte: ByteArray? = null
 
-    constructor(nome: String, numero: Int) {
+    constructor(nome: String, numero: Long) {
         this.nome = nome
         this.numero = numero
     }
-    constructor(id: Int, nome: String, numero: Int) : this(nome, numero){
+    constructor(id: Int, nome: String, numero: Long) : this(nome, numero){
         this.id = id
     }
 
-    constructor(id: Int, nome: String, numero: Int, imageArrayByte: ByteArray) : this(id, nome, numero){
+    constructor(id: Int, nome: String, numero: Long, imageArrayByte: ByteArray) : this(id, nome, numero){
         this.imageArrayByte = imageArrayByte
     }
